:sparkles: feat: Add repository module and seperate model class
- chat class 분리
- repository module 작성
