+++ backend/src/repository/chat.repository.ts
@@ -2,23 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { Repository } from './repository';
 import { nanoid } from 'nanoid';
 
-export type ChatMode = 'private' | 'public' | 'protected';
-export type ChatRole = 'owner' | 'admin' | 'common';
-
-export class Chat {
-  name: string;
-  userList: {
-    id: number;
-    nickname: string;
-    image: string;
-    isAdmin: boolean;
-    isMuted: boolean;
-    isPlayer: boolean;
-  }[];
-  mode: ChatMode = 'public';
-  password?: string;
-  bannedUserIdList?: number[];
-}
+import { Chat } from './model/chat';
 
 @Injectable()
 export class ChatRepository implements Repository<Chat, string> {
