+++ backend/src/repository/model/chat.ts
@@ -1,0 +1,17 @@
+export type ChatMode = 'private' | 'public' | 'protected';
+export type ChatRole = 'owner' | 'admin' | 'common';
+
+export class Chat {
+  name: string;
+  userList: {
+    id: number;
+    nickname: string;
+    image: string;
+    isAdmin: boolean;
+    isMuted: boolean;
+    isPlayer: boolean;
+  }[];
+  mode: ChatMode = 'public';
+  password?: string;
+  bannedUserIdList?: number[];
+}
