+++ backend/src/repository/repository.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { ChatRepository } from './chat.repository';
+
+@Module({
+  providers: [ChatRepository],
+  exports: [ChatRepository],
+})
+export class FriendModule {}
